I *think* I got the multityped domain stuff basically sorted (minus macros, of course)
The only mildly negative side effect was that the domain type needs to be specified *somewhere* explicitly when you're setting up goals and such. On the other hand, I'm not really sure how it was magically picking up OfOne before. I think It'll be totally fine and make sense to just slap a Goal<MyDomain> in at least one spot.