Implemented the member goal function in a new canrun_collections crate
In the process I exposed a flaw in how I was setting up the domains and their traits. Implementing the UnifyIn trait is still rather tricky, and I'm not sure how much value there is in the way I've got it right now. The way that it relates to the domain is a bit funky. It also means that I need to implement the trait for as many primitive types (and collections?) as makes sense or it will be a bit annoying to use.