Got custom unification workng!
I'm not 100% happy with cloning the resolved Rc to do the compare, but it's the cheapest way I can see to give the unify_resolved function the ability to modify the state. I could optimize for the simpler case by passing a borrowed state and making the unify function clone it to return an updated value. This clone would still not be terribly expensive, but I'm not sure if that's the route I want to go either.

More messily, I'm not crazy about how the Unify/DomainType traits are working out, especially the fact that they are both needed in many cases. I'd really like to see if I can clean that up a bit.