Simplify State.unify to only take actual Val<T> instead of allowing the IntoVal
This seems to make a few type inference cases more robust, at the cost of having to do more manual wrapping when calling it. However, most usage is intended to go through the Goal system, where we can still do this conversion. In effect we were doing it at multiple levels. We should really only do it at the boundaries, both to simplify the type checkers job AND to reduce unneeded operations.