Delete the fatally flawed not goal
This should succeed: `both(not(unify(x, 2)), unify(x, 1))`
It doesn't because it can only take the currently resolved state into account.

I think a not_unify goal based on watch that can only compare values directly could be doable, but correct general goal level negation seems difficult. I think it would require the ability to know what variables of a goal tree are not yet resolved. It would then watch those variables and keep evaluating the goal until it is proven false OR all of the vars are resolved.