A state is no longer considered resolved if it has any unsatisfied constraints
I think a more fine-grained approach where we track if a variable relates to any outstanding constraints might be possible, but I'm not sure how tricky it will be, and I'm not sure if it actually makes sense. This is a simple check to implement and seems like the safe bet for now.