Switch State::fork to act on an object implementing the Fork trait
This creates a bit more work for doing custom fork operations, but in practice dealing with complicated closures could get kinda clunky anyway. This also seems to improve the fork heavy benchmarks by ~10%!

The simplification (and possible cause of the performance improvement) is allowing custom forking goals to not require two levels of Rc<Fn>, since a goal specific struct can be created.